// exercices/1.5/mymovies/types.ts
interface Pizza {
  id: number;
  title: string;
  content: string;
}

interface PizzaToUpdate {
  title?: string;
  content?: string;
}

interface Film {
  id: number;
  title: string;
  director: string;
  duration: number; 
  budget?: number;
  description?: string;
  imageUrl?: string;
}

export type { Film };


type NewPizza = Omit<Pizza, "id">;

export type { Pizza, NewPizza, PizzaToUpdate};